fn/requirements.txt
pulsar-client==2.10.1
requests
python-dotenv